pyproject.toml
[project]
name = "meilisearch-admin"
version = "1.0.0"
description = "A web-based administration interface for Meilisearch"
authors = [{ name = "Developer" }]
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "gunicorn>=21.0.0",
    "gevent>=23.0.0",
    "sqlalchemy>=2.0.0",
    "meilisearch>=0.31.0",
    "pyyaml>=6.0",
    "logbook>=1.7.0",
    "simplejson>=3.19.0",
    "marshmallow>=3.20.0",
    "requests>=2.31.0",
    "urllib3>=2.0.0",
    "python-dotenv>=1.0.0",
    "supervisor>=4.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
